fix: widget page and topic page
